# documentation/swagger-api.yaml
swagger: "2.0"
info:
  description: "EPRO Projekt WS2021"
  version: "1.0.0"
  title: "EPRO OKR Projekt"
host: "localhost:8080"
basePath: "/api/v1"
schemes:
  - "http"
paths:
...
  /roles/{id}:
    get:
      tags:
        - "Roles"
      summary: "Get a Role"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of Role to return"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successfully returned role"
          schema:
            $ref: "#/definitions/Roles"
    put:
      tags:
        - "Roles"
      summary: "Update a Role"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the Role to update"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "The updated Role object"
          required: true
          schema:
            $ref: "#/definitions/Role"
      responses:
        200:
          description: "Successfully updated the Role"
    delete:
      tags:
        - "Roles"
      summary: "Deletes a Role"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the Role to be deleted"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successfully deleted the Role"
        404:
          description: "Did not find Role"
...
  /privileges/{id}:
    get:
      tags:
        - "Privileges"
      summary: "Get a Privilege"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of Privilege to return"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successfully returned privilege"
          schema:
            $ref: "#/definitions/Privilege"
    put:
      tags:
        - "Privileges"
      summary: "Update a Privilege"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the Privilege to update"
          required: true
          type: "integer"
        - in: "body"
          name: "body"
          description: "The updated Privilege object"
          required: true
          schema:
            $ref: "#/definitions/Privilege"
      responses:
        200:
          description: "Successfully updated the Privilege"
    delete:
      tags:
        - "Privileges"
      summary: "Deletes a Privilege"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of the Privilege to be deleted"
          required: true
          type: "integer"
      responses:
        200:
          description: "Successfully deleted the Privilege"
        404:
          description: "Did not find Privilege"
...
definitions:
...
  Role:
    type: "object"
    properties:
     id:
      type: "integer"
     name:
      type: "string"
     privilegeId:
        type: "integer"
  Privilege:
    type: "object"
    properties:
      id:
        type: "integer"
      name:
        type: "string"